Revert "mod: int -> ll"
This reverts commit 6ed74231b7fca249a81661dbd89f6bd6989ee342.

+++ library/math/n_choose_k/n_choose_k.hpp
@@ -1,6 +1,6 @@
 /** @file */
 #pragma once
-const ll mod = 17; /**< must be prime */
+const int mod = 17; /**< must be prime */
 vector<ll> inv(2, 1), fact(inv), inv_fact(inv);
 /**
  * @param n,k integers with n < mod
